python/requirements.txt
rich
tqdm
py3createtorrent
unidecode
ffmpeg-python
pathvalidate
pyperclip
chardet
deflacue
Pillow
typer
gspread
pylast
google-auth
